requirements.txt
pygame-ce
pyinstrument # for profiling
numpy 